requirements.txt
torch>=2.0.0
torchvision
pyyaml
numpy
opencv-python
submitit
wandb 